Receiving Follow request from other instances (#154)
✨ Summary
This PR improves the handling of inbound "Follow" requests in the
followerHandler method by:
Ensuring the request type is matched case-sensitively with "Follow".
Improving the error message when the follower account cannot be
retrieved.
Returning more meaningful exceptions when the request type is
unsupported.
Sending an accept Response to the Follower, to notify them that their
request is accepted.

✅ Testing
Tested inbound federated Follow requests from a remote server (e.g.,
Mastodon).
Verified logging of incoming types and error messages.

---------

Co-authored-by: VamshiRajNagula <vamshiraj.nagula@sjsu.edu>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>